chore: update editor toolbar
