+++ src/components/TextEditorBar.jsx
@@ -1,87 +1,56 @@
+import { Button } from "@mantine/core";
 import {
-  TextBoldIcon,
-  TextItaticIcon,
-  TextUnderlineIcon,
-  TextStrikeIcon,
-  TextAlignLeftIcon,
-  TextAlignRightIcon,
-  TextAlignCenterIcon,
-  TextJustifyIcon,
-  TextListIcon,
-  TextListNumberIcon,
-  TextRightIcon,
-  TextLeftIcon,
-} from "@/icons";
-import { Flex } from "@mantine/core";
-import { useMemo } from "react";
+  IconBold,
+  IconItalic,
+  IconUnderline,
+  IconStrikethrough,
+  IconSubscript,
+  IconSuperscript,
+  IconAlignLeft,
+  IconAlignCenter,
+  IconAlignRight,
+  IconAlignJustified,
+  IconIndentIncrease,
+  IconIndentDecrease,
+  IconList,
+  IconListNumbers,
+} from "@tabler/icons-react";
 
 export const TextEditorBar = ({ size = "md" }) => {
-  const Tool1 = [
-    { icon: TextBoldIcon },
-    { icon: TextItaticIcon },
-    { icon: TextUnderlineIcon },
-    { icon: TextStrikeIcon },
-    { text: "x²" },
-    { text: "x₂" },
-  ];
-
-  const Tool2 = [
-    { icon: TextAlignLeftIcon },
-    { icon: TextJustifyIcon },
-    { icon: TextAlignRightIcon },
-    { icon: TextAlignCenterIcon },
-  ];
-
-  const Tool3 = [{ icon: TextRightIcon }, { icon: TextLeftIcon }];
-
-  const Tool4 = [{ icon: TextListIcon }, { icon: TextListNumberIcon }];
-
-  const toolBarSize = useMemo(() => {
-    switch (size) {
-      case "xl":
-        return "w-[4.5rem] h-14";
-      default:
-        return "w-14 h-12";
-    }
-  });
-  const commonStyle = `${toolBarSize} border-2 border-gray-200 justify-center items-center bg-white hover:bg-gray-50 cursor-pointer border-r-0 last:border-r-2`;
+  const EditorIcons = {
+    editor: [
+      <IconBold />,
+      <IconItalic />,
+      <IconUnderline />,
+      <IconStrikethrough />,
+      <IconSubscript />,
+      <IconSuperscript />,
+    ],
+    align: [
+      <IconAlignLeft />,
+      <IconAlignCenter />,
+      <IconAlignRight />,
+      <IconAlignJustified />,
+    ],
+    indent: [<IconIndentIncrease />, <IconIndentDecrease />],
+    list: [<IconList />, <IconListNumbers />],
+  };
 
   return (
-    <Flex className="justify-between mb-2 border-collapse w-full">
-      <Flex>
-        {Tool1.map((item, index) => (
-          <Flex className={commonStyle} key={`tool1_${index}`}>
-            {item.icon ? (
-              <item.icon size={20} />
-            ) : (
-              <p className="text-lg">{item.text}</p>
-            )}
-          </Flex>
-        ))}
-      </Flex>
-
-      <Flex className="ml-2">
-        {Tool2.map((Item, index) => (
-          <Flex className={commonStyle} key={`tool2_${index}`}>
-            <Item.icon size={24} />
-          </Flex>
-        ))}
-      </Flex>
-
-      <Flex className="ml-2">
-        {Tool3.map((Item, index) => (
-          <Flex className={commonStyle} key={`tool3_${index}`}>
-            <Item.icon size={index === 0 ? 18 : 22} />
-          </Flex>
-        ))}
-      </Flex>
-      <Flex className="ml-2">
-        {Tool4.map((Item, index) => (
-          <Flex className={commonStyle} key={`tool4_${index}`}>
-            <Item.icon size={24} />
-          </Flex>
-        ))}
-      </Flex>
-    </Flex>
+    <div className="row justify-between mb-4">
+      {Object.keys(EditorIcons).map((key) => (
+        <Button.Group key={key}>
+          {EditorIcons[key].map((icon, idx) => (
+            <Button
+              key={idx}
+              className="!rounded-none px-3 w-14 h-10"
+              variant="default"
+            >
+              {icon}
+            </Button>
+          ))}
+        </Button.Group>
+      ))}
+    </div>
   );
 };
