+++ src/features/template/components/DisasterComponentList.jsx
@@ -1,46 +1,10 @@
 "use client";
 import { useContext } from "react";
 import clsx from "clsx";
-import { Button, Flex, Image, Textarea } from "@mantine/core";
+import { Flex, Image, Textarea } from "@mantine/core";
 import { CreateLayoutContext } from "@/contexts/CreateLayoutContext";
 import { TextEditorBar, UploadFile } from "@/components";
 import { SPECIAL_TYPE } from "@/constants";
-
-import {
-  IconBold,
-  IconItalic,
-  IconUnderline,
-  IconStrikethrough,
-  IconSubscript,
-  IconSuperscript,
-  IconAlignLeft,
-  IconAlignCenter,
-  IconAlignRight,
-  IconAlignJustified,
-  IconIndentIncrease,
-  IconIndentDecrease,
-  IconList,
-  IconListNumbers,
-} from "@tabler/icons-react";
-
-const EditorIcons = {
-  editor: [
-    <IconBold />,
-    <IconItalic />,
-    <IconUnderline />,
-    <IconStrikethrough />,
-    <IconSubscript />,
-    <IconSuperscript />,
-  ],
-  align: [
-    <IconAlignLeft />,
-    <IconAlignCenter />,
-    <IconAlignRight />,
-    <IconAlignJustified />,
-  ],
-  indent: [<IconIndentIncrease />, <IconIndentDecrease />],
-  list: [<IconList />, <IconListNumbers />],
-};
 
 export const DisasterComponentList = ({ option, components }) => {
   const text = option?.value;
@@ -55,21 +19,7 @@ export const DisasterComponentList = ({ option, components }) => {
             <>
               <p className="text-sm text-gray-400">ข้อความ</p>
               <div className="w-full">
-                <div className="row gap-4 mb-4">
-                  {Object.keys(EditorIcons).map((key) => (
-                    <Button.Group key={key}>
-                      {EditorIcons[key].map((icon, idx) => (
-                        <Button
-                          key={idx}
-                          className="!rounded-none px-3"
-                          variant="default"
-                        >
-                          {icon}
-                        </Button>
-                      ))}
-                    </Button.Group>
-                  ))}
-                </div>
+                <TextEditorBar />
                 <Textarea
                   className="border-0 rounded-none"
                   autosize
