--- src/icons/TextAlignRightIcon.jsx
@@ -1,55 +1,0 @@
-export const TextAlignRightIcon = ({ size, style, fill, ...others }) => {
-  return (
-    <svg
-      width="60px"
-      height="60px"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      style={{ width: size, height: size, ...style }}
-      {...others}
-    >
-      <g id="Complete">
-        <g id="align-right">
-          <g>
-            <polygon
-              fill="#ffffff"
-              points="19.9 18 11.1 18 11.1 18 19.9 18 19.9 18"
-              stroke="#000000"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="2"
-            />
-
-            <polygon
-              fill="#ffffff"
-              points="20 14 4 14 4 14 20 14 20 14"
-              stroke="#000000"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="2"
-            />
-
-            <polygon
-              fill="#ffffff"
-              points="19.9 10 11.1 10 11.1 10 19.9 10 19.9 10"
-              stroke="#000000"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="2"
-            />
-
-            <polygon
-              fill="#ffffff"
-              points="20 6 4 6 4 6 20 6 20 6"
-              stroke="#000000"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="2"
-            />
-          </g>
-        </g>
-      </g>
-    </svg>
-  );
-};
