--- src/icons/TextBoldIcon.jsx
@@ -1,20 +1,0 @@
-export const TextBoldIcon = ({ size, style, fill, ...others }) => {
-  return (
-    <svg
-      fill="#000000"
-      width="30px"
-      height="30px"
-      viewBox="0 0 24 24"
-      id="text-bold"
-      data-name="Flat Color"
-      xmlns="http://www.w3.org/2000/svg"
-      style={{ width: size, height: size, ...style }}
-      {...others}
-    >
-      <path
-        id="primary"
-        d="M16.42,11.35A5.45,5.45,0,0,0,18,7.5,5.51,5.51,0,0,0,12.5,2H5A1,1,0,0,0,5,4H6V20H5a1,1,0,0,0,0,2h9.5a5.5,5.5,0,0,0,1.92-10.65ZM12.5,4a3.5,3.5,0,0,1,0,7H9V4Zm2,16H9V13h5.5a3.5,3.5,0,0,1,0,7Z"
-      ></path>
-    </svg>
-  );
-};
