--- src/icons/TextItaticIcon.jsx
@@ -1,35 +1,0 @@
-export const TextItaticIcon = ({ size, style, fill, ...others }) => {
-  return (
-    <svg
-      width="50px"
-      height="50px"
-      viewBox="0 0 48 48"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      style={{ width: size, height: size, ...style }}
-      {...others}
-    >
-      <path
-        d="M20 6H36"
-        stroke="#000000"
-        strokeWidth="4"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M12 42H28"
-        stroke="#000000"
-        strokeWidth="4"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M29 5.95239L19 42"
-        stroke="#000000"
-        strokeWidth="4"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
