--- src/icons/TextLeftIcon.jsx
@@ -1,44 +1,0 @@
-export const TextLeftIcon = ({ size, style, fill, ...others }) => {
-  return (
-    <svg
-      fill="#000000"
-      id="Capa_1"
-      xmlns="http://www.w3.org/2000/svg"
-      width="50px"
-      height="50px"
-      viewBox="0 0 24 24"
-      style={{ width: size, height: size, ...style }}
-      {...others}
-    >
-      <g
-        id="页面-1"
-        stroke="none"
-        strokeWidth="1"
-        fill="none"
-        fillRule="evenodd"
-      >
-        <g
-          id="Editor"
-          transform="translate(-576.000000, 0.000000)"
-          fillRule="nonzero"
-        >
-          <g
-            id="indent_decrease_line"
-            transform="translate(576.000000, 0.000000)"
-          >
-            <path
-              d="M24,0 L24,24 L0,24 L0,0 L24,0 Z M12.5934901,23.257841 L12.5819402,23.2595131 L12.5108777,23.2950439 L12.4918791,23.2987469 L12.4918791,23.2987469 L12.4767152,23.2950439 L12.4056548,23.2595131 C12.3958229,23.2563662 12.3870493,23.2590235 12.3821421,23.2649074 L12.3780323,23.275831 L12.360941,23.7031097 L12.3658947,23.7234994 L12.3769048,23.7357139 L12.4804777,23.8096931 L12.4953491,23.8136134 L12.4953491,23.8136134 L12.5071152,23.8096931 L12.6106902,23.7357139 L12.6232938,23.7196733 L12.6232938,23.7196733 L12.6266527,23.7031097 L12.609561,23.275831 C12.6075724,23.2657013 12.6010112,23.2592993 12.5934901,23.257841 L12.5934901,23.257841 Z M12.8583906,23.1452862 L12.8445485,23.1473072 L12.6598443,23.2396597 L12.6498822,23.2499052 L12.6498822,23.2499052 L12.6471943,23.2611114 L12.6650943,23.6906389 L12.6699349,23.7034178 L12.6699349,23.7034178 L12.678386,23.7104931 L12.8793402,23.8032389 C12.8914285,23.8068999 12.9022333,23.8029875 12.9078286,23.7952264 L12.9118235,23.7811639 L12.8776777,23.1665331 C12.8752882,23.1545897 12.8674102,23.1470016 12.8583906,23.1452862 L12.8583906,23.1452862 Z M12.1430473,23.1473072 C12.1332178,23.1423925 12.1221763,23.1452606 12.1156365,23.1525954 L12.1099173,23.1665331 L12.0757714,23.7811639 C12.0751323,23.7926639 12.0828099,23.8018602 12.0926481,23.8045676 L12.108256,23.8032389 L12.3092106,23.7104931 L12.3186497,23.7024347 L12.3186497,23.7024347 L12.3225043,23.6906389 L12.340401,23.2611114 L12.337245,23.2485176 L12.337245,23.2485176 L12.3277531,23.2396597 L12.1430473,23.1473072 Z"
-              id="MingCute"
-              fillRule="nonzero"
-            ></path>
-            <path
-              d="M20,18 C20.5523,18 21,18.4477 21,19 C21,19.5523 20.5523,20 20,20 L4,20 C3.44772,20 3,19.5523 3,19 C3,18.4477 3.44772,18 4,18 L20,18 Z M20,13 C20.5523,13 21,13.4477 21,14 C21,14.51285 20.613973,14.9355092 20.1166239,14.9932725 L20,15 L11,15 C10.4477,15 10,14.5523 10,14 C10,13.48715 10.386027,13.0644908 10.8833761,13.0067275 L11,13 L20,13 Z M7.06481,9.17279 C7.3223,9.04127 7.62081,9.2354 7.64813,9.55205 L7.69021121,10.1179975 L7.69021121,10.1179975 L7.71561875,10.5818188 L7.71561875,10.5818188 L7.73546669,11.10897 C7.73814746,11.2018257 7.74047364,11.2971098 7.74238376,11.3947172 L7.74836,12.0074 L7.74836,12.0074 L7.74301235,12.5870197 L7.74301235,12.5870197 L7.72862278,13.11726 L7.72862278,13.11726 L7.70767134,13.5938654 L7.70767134,13.5938654 L7.66936583,14.1988993 L7.66936583,14.1988993 L7.64286,14.5228 L7.64286,14.5228 C7.61644,14.8194 7.34428,14.9843 7.09085,14.8552 L6.80186688,14.7044728 L6.80186688,14.7044728 L6.45586264,14.5154224 L6.45586264,14.5154224 L6.05778896,14.2872856 C5.98737932,14.2459632 5.91500644,14.2029804 5.84077348,14.1583213 L5.37388004,13.8701239 L5.37388004,13.8701239 L4.87802717,13.5480558 L4.87802717,13.5480558 L4.42244719,13.2370606 C4.35091212,13.1870338 4.28160076,13.1379824 4.21455248,13.0900152 L3.83973472,12.8160928 L3.83973472,12.8160928 L3.38348053,12.4654222 C3.3400473,12.431017 3.299074,12.39824 3.2606,12.3672 C3.03479,12.1851 3.02537,11.838 3.24013,11.664 L3.50185728,11.4559336 L3.50185728,11.4559336 L3.82364264,11.2098368 L3.82364264,11.2098368 L4.20299536,10.9316232 L4.20299536,10.9316232 L4.63742472,10.6272064 C4.7911372,10.522024 4.953614,10.41346 5.12444,10.3025 L5.36676407,10.1471309 L5.36676407,10.1471309 L5.82374549,9.8650739 L5.82374549,9.8650739 L6.23952875,9.6208625 L6.23952875,9.6208625 L6.60894281,9.4140215 L6.60894281,9.4140215 L7.06481,9.17279 L7.06481,9.17279 Z M20,8 C20.5523,8 21,8.44772 21,9 C21,9.55228 20.5523,10 20,10 L11,10 C10.4477,10 10,9.55228 10,9 C10,8.44772 10.4477,8 11,8 L20,8 Z M20,3 C20.5523,3 21,3.44772 21,4 C21,4.55228 20.5523,5 20,5 L4,5 C3.44772,5 3,4.55228 3,4 C3,3.44772 3.44772,3 4,3 L20,3 Z"
-              id="形状"
-              fill="#09244B"
-            ></path>
-          </g>
-        </g>
-      </g>
-    </svg>
-  );
-};
