--- src/icons/TextListNumberIcon.jsx
@@ -1,15 +1,0 @@
-export const TextListNumberIcon = ({ size, style, fill, ...others }) => {
-  return (
-    <svg
-      width="50px"
-      height="50px"
-      viewBox="0 0 56 56"
-      fill="#000000"
-      xmlns="http://www.w3.org/2000/svg"
-      style={{ width: size, height: size, ...style }}
-      {...others}
-    >
-      <path d="M 9.4960 19.0117 C 10.4101 19.0117 11.0195 18.4961 11.0195 17.4414 L 11.0195 9.8711 C 11.0195 8.8867 10.3164 8.2539 9.3086 8.2539 C 8.4648 8.2539 7.9492 8.5351 7.3632 8.9336 L 5.7226 10.0820 C 5.2304 10.4102 4.9726 10.7148 4.9726 11.2070 C 4.9726 11.8164 5.4648 12.2383 6.0039 12.2383 C 6.2851 12.2383 6.4492 12.1914 6.8476 11.9102 L 7.9257 11.1836 L 7.9492 11.1836 L 7.9492 17.4414 C 7.9492 18.4961 8.5820 19.0117 9.4960 19.0117 Z M 17.6289 16.1758 L 49.7149 16.1758 C 50.7930 16.1758 51.6133 15.3555 51.6133 14.3008 C 51.6133 13.2226 50.7930 12.4024 49.7149 12.4024 L 17.6289 12.4024 C 16.5742 12.4024 15.7539 13.2226 15.7539 14.3008 C 15.7539 15.3555 16.5742 16.1758 17.6289 16.1758 Z M 5.7226 33.1445 L 11.7460 33.1445 C 12.4023 33.1445 12.8711 32.6992 12.8711 32.0430 C 12.8711 31.3398 12.4023 30.8945 11.7460 30.8945 L 8.3945 30.8945 L 8.3945 30.8242 L 10.3164 29.2773 C 11.9335 27.9648 12.5429 27.2148 12.5429 25.8320 C 12.5429 23.9570 10.9726 22.6914 8.4413 22.6914 C 6.2148 22.6914 4.5742 23.8633 4.5742 25.3633 C 4.5742 26.1133 5.0664 26.5117 5.8632 26.5117 C 6.4023 26.5117 6.7773 26.3477 7.0820 25.8086 C 7.4101 25.2461 7.8554 24.9414 8.4882 24.9414 C 9.1913 24.9414 9.6601 25.3867 9.6601 26.0430 C 9.6601 26.6055 9.3789 27.0977 8.1835 28.0586 L 5.0898 30.5898 C 4.6445 30.9648 4.4570 31.3867 4.4570 31.9024 C 4.4570 32.6289 4.9492 33.1445 5.7226 33.1445 Z M 17.6289 30.3555 L 49.7149 30.3555 C 50.7930 30.3555 51.6133 29.5117 51.6133 28.4570 C 51.6133 27.4024 50.7930 26.5820 49.7149 26.5820 L 17.6289 26.5820 C 16.5742 26.5820 15.7539 27.4024 15.7539 28.4570 C 15.7539 29.5117 16.5742 30.3555 17.6289 30.3555 Z M 8.5117 47.7461 C 11.2773 47.7461 12.9413 46.5039 12.9413 44.5352 C 12.9413 43.2461 12.0273 42.3555 10.4101 42.2148 L 10.4101 42.1445 C 11.6054 41.9336 12.5195 41.1367 12.5195 39.7773 C 12.5195 37.9961 10.7851 37.0117 8.4882 37.0117 C 6.6367 37.0117 4.6445 37.8789 4.6445 39.4258 C 4.6445 40.0820 5.1132 40.5508 5.8398 40.5508 C 6.3554 40.5508 6.5898 40.3398 6.9179 39.9883 C 7.4570 39.4024 7.9023 39.2148 8.4648 39.2148 C 9.1913 39.2148 9.7304 39.5664 9.7304 40.2695 C 9.7304 40.9258 9.1913 41.2539 8.2304 41.2539 L 7.9492 41.2539 C 7.2929 41.2539 6.8711 41.5820 6.8711 42.2617 C 6.8711 42.8945 7.2695 43.2695 7.9492 43.2695 L 8.2539 43.2695 C 9.3086 43.2695 9.8711 43.6211 9.8711 44.3477 C 9.8711 44.9805 9.2851 45.4492 8.5117 45.4492 C 7.6679 45.4492 7.0820 44.9805 6.6367 44.5352 C 6.3554 44.2773 6.1211 44.0898 5.6992 44.0898 C 4.9257 44.0898 4.3867 44.5352 4.3867 45.2617 C 4.3867 46.8789 6.5898 47.7461 8.5117 47.7461 Z M 17.6289 44.5117 L 49.7149 44.5117 C 50.7930 44.5117 51.6133 43.6914 51.6133 42.6367 C 51.6133 41.5586 50.7930 40.7383 49.7149 40.7383 L 17.6289 40.7383 C 16.5742 40.7383 15.7539 41.5586 15.7539 42.6367 C 15.7539 43.6914 16.5742 44.5117 17.6289 44.5117 Z" />
-    </svg>
-  );
-};
