--- src/icons/TextRightIcon.jsx
@@ -1,39 +1,0 @@
-export const TextRightIcon = ({ size, style, fill, ...others }) => {
-  return (
-    <svg
-      fill="#000000"
-      id="Capa_1"
-      xmlns="http://www.w3.org/2000/svg"
-      width="50px"
-      height="50px"
-      viewBox="0 0 508.865 508.866"
-      style={{ width: size, height: size, ...style }}
-      {...others}
-    >
-      <g>
-        <g>
-          <path
-            d="M476.992,43.014H31.873C14.296,43.014,0,57.311,0,74.887s14.296,31.873,31.873,31.873h445.119
-			c17.578,0,31.873-14.296,31.873-31.873S494.57,43.014,476.992,43.014z"
-          />
-          <path
-            d="M476.992,160.488h-218.68c-17.576,0-31.872,14.296-31.872,31.873s14.296,31.873,31.872,31.873h218.68
-			c17.578,0,31.873-14.296,31.873-31.873S494.57,160.488,476.992,160.488z"
-          />
-          <path
-            d="M476.992,277.955h-218.68c-17.576,0-31.872,14.297-31.872,31.873c0,17.577,14.296,31.873,31.872,31.873h218.68
-			c17.578,0,31.873-14.296,31.873-31.873C508.865,292.251,494.57,277.955,476.992,277.955z"
-          />
-          <path
-            d="M476.992,402.105H31.873C14.296,402.105,0,416.402,0,433.978c0,17.577,14.296,31.873,31.873,31.873h445.119
-			c17.578,0,31.873-14.296,31.873-31.873C508.865,416.402,494.57,402.105,476.992,402.105z"
-          />
-          <path
-            d="M30.729,351.884l149.249-85.153c11.903-6.793,11.903-17.803,0-24.596L30.729,156.981
-			c-11.897-6.793-21.548-1.187-21.548,12.516V339.37C9.18,353.072,18.832,358.678,30.729,351.884z"
-          />
-        </g>
-      </g>
-    </svg>
-  );
-};
