--- src/icons/TextStrikeIcon.jsx
@@ -1,42 +1,0 @@
-export const TextStrikeIcon = ({ size, style, fill, ...others }) => {
-  return (
-    <svg
-      width="50px"
-      height="50px"
-      viewBox="0 0 48 48"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      style={{ width: size, height: size, ...style }}
-      {...others}
-    >
-      <path
-        d="M5 24H43"
-        stroke="#000000"
-        strokeWidth="4"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M24 24C40 30 34 44 24 44C13.9999 44 12 36 12 36"
-        stroke="#000000"
-        strokeWidth="4"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M35.9999 12C35.9999 12 33 4 23.9999 4C14.9999 4 11.4359 11.5995 15.6096 18"
-        stroke="#000000"
-        strokeWidth="4"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M12 36C12 36 15.9999 44 24 44C32 44 36.564 36.4005 32.3903 30"
-        stroke="#000000"
-        strokeWidth="4"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
