--- src/icons/TextUnderlineIcon.jsx
@@ -1,22 +1,0 @@
-export const TextUnderlineIcon = ({ size, style, fill, ...others }) => {
-  return (
-    <svg
-      width="50px"
-      height="50px"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      style={{ width: size, height: size, ...style }}
-      {...others}
-    >
-      <path
-        d="M7 3C7 2.44772 6.55228 2 6 2C5.44772 2 5 2.44772 5 3V10C5 11.8565 5.7375 13.637 7.05025 14.9497C8.36301 16.2625 10.1435 17 12 17C13.8565 17 15.637 16.2625 16.9497 14.9497C18.2625 13.637 19 11.8565 19 10V3C19 2.44772 18.5523 2 18 2C17.4477 2 17 2.44772 17 3V10C17 11.3261 16.4732 12.5979 15.5355 13.5355C14.5979 14.4732 13.3261 15 12 15C10.6739 15 9.40215 14.4732 8.46447 13.5355C7.52678 12.5979 7 11.3261 7 10V3Z"
-        fill="#000000"
-      />
-      <path
-        d="M4 20C3.44772 20 3 20.4477 3 21C3 21.5523 3.44772 22 4 22H20C20.5523 22 21 21.5523 21 21C21 20.4477 20.5523 20 20 20H4Z"
-        fill="#000000"
-      />
-    </svg>
-  );
-};
