+++ src/icons/index.js
@@ -40,15 +40,3 @@ export * from "./Pdf";
 export * from "./Png";
 export * from "./History";
 export * from "./Sent";
-export * from "./TextBoldIcon";
-export * from "./TextItaticIcon";
-export * from "./TextUnderlineIcon";
-export * from "./TextStrikeIcon";
-export * from "./TextAlignLeftIcon";
-export * from "./TextAlignRightIcon";
-export * from "./TextAlignCenterIcon";
-export * from "./TextListIcon";
-export * from "./TextListNumberIcon";
-export * from "./TextJustifyIcon";
-export * from "./TextRightIcon";
-export * from "./TextLeftIcon";
